+++ src/App.js
@@ -3,9 +3,10 @@ import './App.css';
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './components/Home';
-import AddGroup from './components/redux/Groups/AddGroup';
+import AddGroup from './components/redux/Groups/GroupReducer';
 import DashBoard from './components/Dashboard';
 import PostGroup from './components/redux/Groups/PostGroup';
+import GetGroups from './components/redux/Groups/GetGroups';
 
 
 function App() {
@@ -17,6 +18,8 @@ function App() {
         <Route exact path="/dashboard" element={<DashBoard />} />
         <Route exact path ="groups/add" element={<AddGroup />} />
         <Route exact path ="/add" element={<PostGroup />} />
+        <Route exact path ="/groups" element={<GetGroups />} />
+
 
 
       </Routes>
